pyproject.toml
[tool.poetry]
name = "question-and-answer-quiz"
version = "0.2.0"
description = "Simple question and answe quiz game"
authors = ["Giovani Liskoski Zanini <giovanilzanini@hotmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "question_and_answer_quiz" }]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.5"
djangorestframework = "^3.14.0"
Markdown = "^3.4.1"
python-dotenv = "^0.21.1"
python-jose = "^3.3.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.1"
pylint = "^2.15.10"
black = "^22.12.0"
pytest-django = "^4.5.2"
pytest-freezegun = "^0.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
